use ulong for scanf compatibility

+++ src/interactive.c
@@ -75,7 +75,7 @@ void write_board_interactive(sized_board* board) {
 tile* choose_tile_interactive(sized_tlist* list, tile** t) {
     tile* temp = *t;
     fputs("choose tile (number): ", stdout);
-    size_t choice;
+    ulong choice;
     while (true) {
         scanf("%lu", &choice);
         // exhaust stdin
@@ -121,15 +121,11 @@ void rotate_tile_interactive(tile** t) {
     }
 }
 
-// zmień x y na h w
-// numerowanie od 1 nie 0
-// naprawic info o rotacji
-// jakaś niezła imba daje rzeczy których nie ma
 void place_tile_interactive(sized_board* board, sized_tlist* list, tile** t) {
     if (*t == 0) {
         choose_tile_interactive(list, t);
     }
-    size_t h, w;
+    ulong h, w;
     while (true) {
         fputs("where to place tile (h w): ", stdout);
         if (scanf("%lu %lu", &h, &w) == 2) {
